pyproject.toml
[project]
name = "rush-roster-field"
version = "0.1.0"
description = "Field device for speed monitoring system"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "sqlalchemy>=2.0.0",
    "pyyaml>=6.0",
    "opencv-python>=4.8.0",
    "pyserial>=3.5",
    "requests>=2.31.0",
    "python-multipart>=0.0.6",
    "jinja2>=3.1.2",
    "cv2-enumerate-cameras>=1.1.4",
    "psutil>=7.1.0",
    "supervision>=0.26.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
